refactor value parsers
