+++ src/Discord.Net.ComponentDesigner.Generator/Nodes/ComponentPropertyValidator.cs
@@ -1,0 +1,7 @@
+﻿namespace Discord.ComponentDesigner.Generator.Nodes;
+
+public delegate void ComponentPropertyValidator<T>(
+    ComponentNode node,
+    ComponentProperty<T> property,
+    ComponentNodeContext context
+);
