+++ src/Discord.Net.ComponentDesigner.Generator/Nodes/ValueParsers/ValueParseDelegate.cs
@@ -1,0 +1,3 @@
+﻿namespace Discord.ComponentDesigner.Generator.Nodes;
+
+public delegate ComponentPropertyValue<T>? ValueParseDelegate<T>(ComponentProperty<T> property);
